F Fix styling og tekst i andreytelser-komponent
