+++ src/components/behandlingSteg/BehandlingSteg.tsx
@@ -4,7 +4,6 @@ import { ReactElement } from 'react';
 import { Behandlingssteg } from '../../types/Behandlingssteg';
 import Alder from '../steg/Alder';
 import styles from './BehandlingSteg.module.css';
-import { AndreYtelser } from '../steg/AndreYtelser';
 import Tiltaksdeltagelse from '../steg/Tiltaksdeltagelse';
 import Stønadsdager from '../steg/Stønadsdager';
 import Kvalifiseringsprogrammet from '../steg/Kvalifiseringsprogrammet';
@@ -12,7 +11,7 @@ import Institusjonsopphold from '../steg/Institusjonsopphold';
 import Introduksjonsprogrammet from '../steg/Introduksjonsprogrammet';
 import Oppsummering from '../steg/Oppsummering';
 import FristForFramsettingAvKrav from '../steg/FristForFramsettingAvKrav';
-import { AndreYtelserMinimalistisk } from '../steg/AndreYtelserMinimalistisk';
+import { AndreYtelser } from '../steg/AndreYtelser';
 
 const BehandlingSteg = () => {
   const router = useRouter();
@@ -33,7 +32,7 @@ const BehandlingSteg = () => {
       case Behandlingssteg.INSTITUSJONSOPPHOLD:
         return <Institusjonsopphold />;
       case Behandlingssteg.ANDREYTELSER:
-        return <AndreYtelserMinimalistisk />;
+        return <AndreYtelser />;
       case Behandlingssteg.STØNADSDAGER:
         return <Stønadsdager />;
       case Behandlingssteg.OPPSUMMERING:
