+++ src/components/steg/AndreYtelser.tsx
@@ -1,59 +1,124 @@
-import { Loader, VStack } from '@navikt/ds-react';
-import { SaksopplysningTabell } from '../saksopplysning-tabell/SaksopplysningTabell';
+import {
+  Alert,
+  Button,
+  Heading,
+  Loader,
+  Radio,
+  RadioGroup,
+  VStack,
+} from '@navikt/ds-react';
 import { useRouter } from 'next/router';
-import { useContext } from 'react';
-import { useHentBehandling } from '../../hooks/useHentBehandling';
-import { SaksbehandlerContext } from '../../pages/_app';
-import { avklarLesevisning } from '../../utils/avklarLesevisning';
 import StegHeader from './StegHeader';
 import UtfallstekstMedIkon from './UtfallstekstMedIkon';
+import { useHentLivsopphold } from '../../hooks/useHentLivsopphold';
+import { useSWRConfig } from 'swr';
+import { Controller, SubmitHandler, useForm } from 'react-hook-form';
+import styles from './Steg.module.css';
+
+export interface SkjemaFelter {
+  harAndreYtelser: boolean;
+}
 
 export const AndreYtelser = () => {
   const router = useRouter();
   const behandlingId = router.query.behandlingId as string;
-  const { valgtBehandling, isLoading } = useHentBehandling(behandlingId);
-  const { innloggetSaksbehandler } = useContext(SaksbehandlerContext);
+  const { livsopphold, isLoading } = useHentLivsopphold(behandlingId);
+  const mutator = useSWRConfig().mutate;
 
-  if (isLoading || !valgtBehandling) {
+  const { handleSubmit, control, watch } = useForm<SkjemaFelter>({
+    mode: 'onSubmit',
+    defaultValues: {
+      harAndreYtelser: undefined,
+    },
+  });
+
+  const watchHarLivsoppholdytelser = watch('harAndreYtelser');
+
+  if (isLoading || !livsopphold) {
     return <Loader />;
   }
 
-  const andreYtelser = valgtBehandling.ytelsessaksopplysninger;
+  const håndterLagreLivsoppholdSaksopplysning = (harYtelser: boolean) => {
+    if (harYtelser) {
+      return;
+    }
 
-  const girInnvilget = valgtBehandling.samletUtfall === 'OPPFYLT';
+    fetch(`/api/behandling/${behandlingId}/vilkar/livsopphold`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        ytelseForPeriode: [
+          {
+            periode: livsopphold.vurderingsPeriode,
+            deltar: harYtelser,
+          },
+        ],
+      }),
+    })
+      .then(() => {
+        mutator(`/api/behandling/${behandlingId}/vilkar/livsopphold`);
+      })
+      .catch((error) => {
+        throw new Error(
+          `Noe gikk galt ved lagring av antall dager: ${error.message}`,
+        );
+      });
+  };
 
-  const lesevisning = avklarLesevisning(
-    innloggetSaksbehandler!!,
-    valgtBehandling.saksbehandler,
-    valgtBehandling.beslutter,
-    valgtBehandling.behandlingsteg,
-    girInnvilget,
-  );
+  const onSubmit: SubmitHandler<SkjemaFelter> = (data) => {
+    håndterLagreLivsoppholdSaksopplysning(data.harAndreYtelser);
+  };
 
-  if (!andreYtelser) return <Loader />;
   return (
     <VStack gap="4">
       <StegHeader
         headertekst={'Forholdet til andre ytelser'}
-        lovdatatekst={andreYtelser.vilkårLovreferanse.beskrivelse}
-        paragraf={andreYtelser.vilkårLovreferanse.paragraf}
+        lovdatatekst={livsopphold.vilkårLovreferanse.beskrivelse}
+        paragraf={livsopphold.vilkårLovreferanse.paragraf}
         lovdatalenke={
           'https://lovdata.no/dokument/SF/forskrift/2013-11-04-1286'
         }
       />
-      <UtfallstekstMedIkon samletUtfall={andreYtelser.samletUtfall} />
-      <SaksopplysningTabell
-        saksopplysninger={andreYtelser.saksopplysninger.filter(
-          (saksopplysning) =>
-            saksopplysning.saksopplysningTittel !=
-              'Kvalifiseringsprogrammet(KVP)' &&
-            saksopplysning.saksopplysningTittel != 'Introduksjonsprogrammet' &&
-            saksopplysning.saksopplysningTittel != 'Institusjonsopphold',
-        )}
-        behandlingId={valgtBehandling.behandlingId}
-        vurderingsperiode={valgtBehandling.vurderingsperiode}
-        lesevisning={lesevisning}
-      />
+      {!watchHarLivsoppholdytelser && (
+        <UtfallstekstMedIkon samletUtfall={livsopphold.samletUtfall} />
+      )}
+      {watchHarLivsoppholdytelser && (
+        <Alert variant="error">
+          <Heading spacing size="small" level="3">
+            Utfall støttes ikke
+          </Heading>
+          Denne vurderingen fører til et utfall vi ikke støtter i denne
+          løsningen enda.
+        </Alert>
+      )}
+      <form onSubmit={handleSubmit(onSubmit)}>
+        <Controller
+          name={'harAndreYtelser'}
+          control={control}
+          render={({ field: { onChange } }) => {
+            return (
+              <RadioGroup
+                legend="Har bruker andre ytelser til livsopphold i vurderingsperioden?"
+                onChange={onChange}
+                defaultValue={undefined}
+              >
+                <Radio value={true}>{`Ja`}</Radio>
+                <Radio value={false}>{`Nei`}</Radio>
+              </RadioGroup>
+            );
+          }}
+        />
+        <Button
+          type="submit"
+          value="submit"
+          className={styles.marginTop}
+          disabled={watchHarLivsoppholdytelser}
+        >
+          Lagre endring
+        </Button>
+      </form>
     </VStack>
   );
 };
