+++ src/components/steg/Steg.module.css
@@ -10,3 +10,7 @@
     margin-left: 2rem;
     margin-right: 2rem;
 }
+
+.marginTop {
+    margin-top: 1rem;
+}
